refactor(llm): ♻️ extract load_prompt to utils.prompt
